==> notebooks/2_EDA.ipynb <==
Data exploration of May pickled data.

import pandas as pd
import pickle
import matplotlib.pyplot as plt
import seaborn as sns

#this is a python magic function to automatically show matplotlib plots
%matplotlib inline  

may = pd.read_pickle(r'../data/may.pkl')

may.head()

may.tail()

may.info()

may.isnull().sum()

may.companyname.value_counts()

company_dict = {'Bird':0, 'Lyft': 1, 'Gotcha': 2, 'Lime': 3, 'Spin': 4, 'Jump': 5, 'Bolt': 6}

So the above info tells us that in May, Lime was the company with the most scooter rides, followed by Bird and then Lyft.

may.describe()

sns.boxplot(y=may.companyname);
