==> notebooks/4_MayScooterRides.ipynb <==
from shapely.geometry import Point
import pandas as pd
import pickle
import matplotlib.pyplot as plt
from datetime import date, timedelta, datetime
import geopandas as gpd

loading may data - pickle and csv


%%time
may_loc = pd.read_pickle("../data/may.pkl")
may_trip = pd.read_csv("../data/may_trip.csv")

may_trip['pubTimeStamp'] = pd.to_datetime(may_trip['pubTimeStamp'])


This section makes a dictionary of dataframes. The dataframes can be "chunked up" into different time segments. The main things you need to adjust: the start and end variables, and the delta variable, which designates how "big" a "time chunk" should be.

#make a list of datetimes with datetime function 
start = datetime(2019, 5, 26, 0, 0)
end = datetime(2019, 5, 27, 0, 0)
delta = timedelta(days=1)
timelist = []
timenamelist = []
while start <= end:
    timelist.append(start)
    timenamelist.append(start.strftime("%Y-%m-%d-%H-%M"))
    start += delta
# print(timelist)
# print(timenamelist)

#goal of function: using list of time frames, in this case, days, make a dataframe containing
#info about the location for each scooter, for each day


#make dictionary of dataframes using base dataframe and a user-defined segments_of_time_list
#from cell above
def make_day_df(datetime_list, timenames_list, dataframe, colname):
    resultslist = {} #turn into dictionary
    x = 0
    y = 1
    z = 0
    while x in range(0,len(datetime_list)-1):
        starttime = datetime_list[x]
        endtime = datetime_list[y]
        mask = (dataframe[colname] > starttime) & (dataframe[colname] <= endtime)
        framename = dataframe.loc[mask]
        x += 1
        y += 1
        print(x)
        resultslist[timenames_list[z]] = framename
        z += 1
    return resultslist
        
the_output = make_day_df(timelist, timenamelist, may_trip,"pubTimeStamp")


the_output.keys()

I'm interested in May 26th specifically, the Saturday of Memorial Day weekend, since it had the most scooter rides.

the_output['2019-05-26-00-00']
# the_output['2019-05-01-00-00']['sumdid'].nunique()

the_output['2019-05-26-00-00']['sumdID'].value_counts()
# the_output['2019-05-01-00-00']['sumdid'].nunique()

What I'd like to do is return only the Scooter IDs that had less than 3 rides on this day. Then be able to plot those points to see where they are clustered, because that would tell us where we can reduce the scooter density.

the_output['2019-05-26-00-00']['sumdID'].value_counts().loc[lambda x : x<3]

So now we can see that on 5/26/19, 3878 (subtracting the row for unknowns) unique scooter IDs had rides. Of those 3878 scooters, 1499 only had 1 or 2 rides (not meeting the 3 rides per day threshold). So only 61% of scooters on this day had 3 or more rides.

I'm also curious to know what percentage of scooters are over-performing though. For instance, how many scooters are being used for at least 6 rides per day (twice the stated goal)?

the_output['2019-05-26-00-00']['sumdID'].value_counts().loc[lambda x : x>5]

So 862 unique scooters had twice (or more) the stated rides-per-day goal on this particular day. That's 22% of the scooters.

May 26th was the busiest day of the month, and the Saturday of a holiday weekend... but what about just a regular weekday? Since the aim of this project is supposed to be about getting scooters used by regular Nashville residents, let's see what the data shows us. For May, the median number of rides per day occured on May 16th. So let's run the same process for that day - what percentage of scooters had fewer than the goal of 3 rides per day, and what percentage of scooters had 6 rides or more (twice the stated goal)?

#make a list of datetimes with datetime function 
start = datetime(2019, 5, 16, 0, 0)
end = datetime(2019, 5, 17, 0, 0)
delta = timedelta(days=1)
timelist = []
timenamelist = []
while start <= end:
    timelist.append(start)
    timenamelist.append(start.strftime("%Y-%m-%d-%H-%M"))
    start += delta
# print(timelist)
# print(timenamelist)

#goal of function: using list of time frames, in this case, days, make a dataframe containing
#info about the location for each scooter, for each day


#make dictionary of dataframes using base dataframe and a user-defined segments_of_time_list
#from cell above
def make_day_df(datetime_list, timenames_list, dataframe, colname):
    resultslist = {} #turn into dictionary
    x = 0
    y = 1
    z = 0
    while x in range(0,len(datetime_list)-1):
        starttime = datetime_list[x]
        endtime = datetime_list[y]
        mask = (dataframe[colname] > starttime) & (dataframe[colname] <= endtime)
        framename = dataframe.loc[mask]
        x += 1
        y += 1
        print(x)
        resultslist[timenames_list[z]] = framename
        z += 1
    return resultslist
        
the_output = make_day_df(timelist, timenamelist, may_trip,"pubTimeStamp")


the_output.keys()

the_output['2019-05-16-00-00']
# the_output['2019-05-01-00-00']['sumdid'].nunique()

the_output['2019-05-16-00-00']['sumdID'].value_counts()
# the_output['2019-05-01-00-00']['sumdid'].nunique()

the_output['2019-05-16-00-00']['sumdID'].value_counts().loc[lambda x : x<3]

So out of 1665 unique scooters that ran on 5/16/19, 1153 of them had fewer than 3 rides. So only 31% of the scooters met the threshold of 3 rides that day.

the_output['2019-05-16-00-00']['sumdID'].value_counts().loc[lambda x : x>5]

137 scooters, or 8%, performed well with at least 6 rides that day.

#make a list of datetimes with datetime function 
start = datetime(2019, 5, 9, 0, 0)
end = datetime(2019, 5, 10, 0, 0)
delta = timedelta(days=1)
timelist = []
timenamelist = []
while start <= end:
    timelist.append(start)
    timenamelist.append(start.strftime("%Y-%m-%d-%H-%M"))
    start += delta
# print(timelist)
# print(timenamelist)

#goal of function: using list of time frames, in this case, days, make a dataframe containing
#info about the location for each scooter, for each day


#make dictionary of dataframes using base dataframe and a user-defined segments_of_time_list
#from cell above
def make_day_df(datetime_list, timenames_list, dataframe, colname):
    resultslist = {} #turn into dictionary
    x = 0
    y = 1
    z = 0
    while x in range(0,len(datetime_list)-1):
        starttime = datetime_list[x]
        endtime = datetime_list[y]
        mask = (dataframe[colname] > starttime) & (dataframe[colname] <= endtime)
        framename = dataframe.loc[mask]
        x += 1
        y += 1
        print(x)
        resultslist[timenames_list[z]] = framename
        z += 1
    return resultslist
        
the_output = make_day_df(timelist, timenamelist, may_trip,"pubTimeStamp")


the_output.keys()

the_output['2019-05-09-00-00']
# the_output['2019-05-01-00-00']['sumdid'].nunique()

the_output['2019-05-09-00-00']['sumdID'].value_counts()
# the_output['2019-05-01-00-00']['sumdid'].nunique()

1737 rides total

the_output['2019-05-09-00-00']['sumdID'].value_counts().loc[lambda x : x<3]

654 scooters had 3 or more rides, or 38%

the_output['2019-05-09-00-00']['sumdID'].value_counts().loc[lambda x : x>5]

129 scooters had 6 or more rides, or 7%
