==> notebooks/6_JuneScooterRides.ipynb <==
from shapely.geometry import Point
import pandas as pd
import pickle
import matplotlib.pyplot as plt
from datetime import date, timedelta, datetime
import geopandas as gpd

loading may data - pickle and csv


%%time
june_loc = pd.read_pickle("../data/june.pkl")
june_trip = pd.read_csv("../data/june_trip.csv")

june_trip['pubTimeStamp'] = pd.to_datetime(june_trip['pubTimeStamp'])


This section makes a dictionary of dataframes. The dataframes can be "chunked up" into different time segments. The main things you need to adjust: the start and end variables, and the delta variable, which designates how "big" a "time chunk" should be.

#make a list of datetimes with datetime function 
start = datetime(2019, 6, 8, 0, 0)
end = datetime(2019, 6, 9, 0, 0)
delta = timedelta(days=1)
timelist = []
timenamelist = []
while start <= end:
    timelist.append(start)
    timenamelist.append(start.strftime("%Y-%m-%d-%H-%M"))
    start += delta
# print(timelist)
# print(timenamelist)

#goal of function: using list of time frames, in this case, days, make a dataframe containing
#info about the location for each scooter, for each day


#make dictionary of dataframes using base dataframe and a user-defined segments_of_time_list
#from cell above
def make_day_df(datetime_list, timenames_list, dataframe, colname):
    resultslist = {} #turn into dictionary
    x = 0
    y = 1
    z = 0
    while x in range(0,len(datetime_list)-1):
        starttime = datetime_list[x]
        endtime = datetime_list[y]
        mask = (dataframe[colname] > starttime) & (dataframe[colname] <= endtime)
        framename = dataframe.loc[mask]
        x += 1
        y += 1
        print(x)
        resultslist[timenames_list[z]] = framename
        z += 1
    return resultslist
        
the_output = make_day_df(timelist, timenamelist, june_trip,"pubTimeStamp")


the_output.keys()

I'm interested in May 26th specifically, the Saturday of Memorial Day weekend, since it had the most scooter rides.

the_output['2019-06-08-00-00']
# the_output['2019-05-01-00-00']['sumdid'].nunique()

the_output['2019-06-08-00-00']['sumdID'].value_counts()
# the_output['2019-05-01-00-00']['sumdid'].nunique()

What I'd like to do is return only the Scooter IDs that had less than 3 rides on this day. Then be able to plot those points to see where they are clustered, because that would tell us where we can reduce the scooter density.

the_output['2019-06-08-00-00']['sumdID'].value_counts().loc[lambda x : x<3]

So now we can see that on 6/8/19, 3549 (subtracting the row for unknowns) unique scooter IDs had rides. Of those 3878 scooters, 1569 only had 1 or 2 rides (not meeting the 3 rides per day threshold). So only 56% of scooters on this day had 3 or more rides.

I'm also curious to know what percentage of scooters are over-performing though. For instance, how many scooters are being used for at least 6 rides per day (twice the stated goal)?

the_output['2019-06-08-00-00']['sumdID'].value_counts().loc[lambda x : x>5]

So 573 unique scooters had twice (or more) the stated rides-per-day goal on this particular day. That's 16% of the scooters.

#make a list of datetimes with datetime function 
start = datetime(2019, 6, 28, 0, 0)
end = datetime(2019, 6, 29, 0, 0)
delta = timedelta(days=1)
timelist = []
timenamelist = []
while start <= end:
    timelist.append(start)
    timenamelist.append(start.strftime("%Y-%m-%d-%H-%M"))
    start += delta
# print(timelist)
# print(timenamelist)

#goal of function: using list of time frames, in this case, days, make a dataframe containing
#info about the location for each scooter, for each day


#make dictionary of dataframes using base dataframe and a user-defined segments_of_time_list
#from cell above
def make_day_df(datetime_list, timenames_list, dataframe, colname):
    resultslist = {} #turn into dictionary
    x = 0
    y = 1
    z = 0
    while x in range(0,len(datetime_list)-1):
        starttime = datetime_list[x]
        endtime = datetime_list[y]
        mask = (dataframe[colname] > starttime) & (dataframe[colname] <= endtime)
        framename = dataframe.loc[mask]
        x += 1
        y += 1
        print(x)
        resultslist[timenames_list[z]] = framename
        z += 1
    return resultslist
        
the_output = make_day_df(timelist, timenamelist, june_trip,"pubTimeStamp")


the_output.keys()

the_output['2019-06-28-00-00']
# the_output['2019-05-01-00-00']['sumdid'].nunique()

the_output['2019-06-28-00-00']['sumdID'].value_counts()
# the_output['2019-05-01-00-00']['sumdid'].nunique()

the_output['2019-06-28-00-00']['sumdID'].value_counts().loc[lambda x : x<3]

So out of 2092 unique scooters that ran on 6/28/19, 1411 of them had fewer than 3 rides. So only 33% of the scooters met the threshold of 3 rides that day.

the_output['2019-06-28-00-00']['sumdID'].value_counts().loc[lambda x : x>5]

193 scooters, or 9%, performed well with at least 6 rides that day.

#make a list of datetimes with datetime function 
start = datetime(2019, 6, 24, 0, 0)
end = datetime(2019, 6, 25, 0, 0)
delta = timedelta(days=1)
timelist = []
timenamelist = []
while start <= end:
    timelist.append(start)
    timenamelist.append(start.strftime("%Y-%m-%d-%H-%M"))
    start += delta
# print(timelist)
# print(timenamelist)

#goal of function: using list of time frames, in this case, days, make a dataframe containing
#info about the location for each scooter, for each day


#make dictionary of dataframes using base dataframe and a user-defined segments_of_time_list
#from cell above
def make_day_df(datetime_list, timenames_list, dataframe, colname):
    resultslist = {} #turn into dictionary
    x = 0
    y = 1
    z = 0
    while x in range(0,len(datetime_list)-1):
        starttime = datetime_list[x]
        endtime = datetime_list[y]
        mask = (dataframe[colname] > starttime) & (dataframe[colname] <= endtime)
        framename = dataframe.loc[mask]
        x += 1
        y += 1
        print(x)
        resultslist[timenames_list[z]] = framename
        z += 1
    return resultslist
        
the_output = make_day_df(timelist, timenamelist, june_trip,"pubTimeStamp")


the_output.keys()

the_output['2019-06-24-00-00']
# the_output['2019-05-01-00-00']['sumdid'].nunique()

the_output['2019-06-24-00-00']['sumdID'].value_counts()
# the_output['2019-05-01-00-00']['sumdid'].nunique()

1829 scooters total

the_output['2019-06-24-00-00']['sumdID'].value_counts().loc[lambda x : x<3]

1438 scooters had less than 3 rides. So only 391 had 3 or more. That's 21%

the_output['2019-06-24-00-00']['sumdID'].value_counts().loc[lambda x : x>5]

45 scooters had 6 or more rides, or 2%
