==> notebooks/8_JulyScooterRides.ipynb <==
from shapely.geometry import Point
import pandas as pd
import pickle
import matplotlib.pyplot as plt
from datetime import date, timedelta, datetime
import geopandas as gpd

loading may data - pickle and csv


%%time
july_loc = pd.read_pickle("../data/july.pkl")
july_trip = pd.read_csv("../data/july_trip.csv")

july_trip['pubTimeStamp'] = pd.to_datetime(july_trip['pubTimeStamp'])


This section makes a dictionary of dataframes. The dataframes can be "chunked up" into different time segments. The main things you need to adjust: the start and end variables, and the delta variable, which designates how "big" a "time chunk" should be.

#make a list of datetimes with datetime function 
start = datetime(2019, 7, 6, 0, 0)
end = datetime(2019, 7, 7, 0, 0)
delta = timedelta(days=1)
timelist = []
timenamelist = []
while start <= end:
    timelist.append(start)
    timenamelist.append(start.strftime("%Y-%m-%d-%H-%M"))
    start += delta
# print(timelist)
# print(timenamelist)

#goal of function: using list of time frames, in this case, days, make a dataframe containing
#info about the location for each scooter, for each day


#make dictionary of dataframes using base dataframe and a user-defined segments_of_time_list
#from cell above
def make_day_df(datetime_list, timenames_list, dataframe, colname):
    resultslist = {} #turn into dictionary
    x = 0
    y = 1
    z = 0
    while x in range(0,len(datetime_list)-1):
        starttime = datetime_list[x]
        endtime = datetime_list[y]
        mask = (dataframe[colname] > starttime) & (dataframe[colname] <= endtime)
        framename = dataframe.loc[mask]
        x += 1
        y += 1
        print(x)
        resultslist[timenames_list[z]] = framename
        z += 1
    return resultslist
        
the_output = make_day_df(timelist, timenamelist, july_trip,"pubTimeStamp")


the_output.keys()

I'm interested in May 26th specifically, the Saturday of Memorial Day weekend, since it had the most scooter rides.

the_output['2019-07-06-00-00']
# the_output['2019-05-01-00-00']['sumdid'].nunique()

the_output['2019-07-06-00-00']['sumdID'].value_counts()
# the_output['2019-05-01-00-00']['sumdid'].nunique()

2630 unique scooters this day (subtracted unknowns)

the_output['2019-07-06-00-00']['sumdID'].value_counts().loc[lambda x : x<3]

1341 had fewer than 3 rides. 1315 had 3 or more, or 50%.

the_output['2019-07-06-00-00']['sumdID'].value_counts().loc[lambda x : x>5]

303 had 6 or more rides. 12%

#make a list of datetimes with datetime function 
start = datetime(2019, 7, 25, 0, 0)
end = datetime(2019, 7, 26, 0, 0)
delta = timedelta(days=1)
timelist = []
timenamelist = []
while start <= end:
    timelist.append(start)
    timenamelist.append(start.strftime("%Y-%m-%d-%H-%M"))
    start += delta
# print(timelist)
# print(timenamelist)

#goal of function: using list of time frames, in this case, days, make a dataframe containing
#info about the location for each scooter, for each day


#make dictionary of dataframes using base dataframe and a user-defined segments_of_time_list
#from cell above
def make_day_df(datetime_list, timenames_list, dataframe, colname):
    resultslist = {} #turn into dictionary
    x = 0
    y = 1
    z = 0
    while x in range(0,len(datetime_list)-1):
        starttime = datetime_list[x]
        endtime = datetime_list[y]
        mask = (dataframe[colname] > starttime) & (dataframe[colname] <= endtime)
        framename = dataframe.loc[mask]
        x += 1
        y += 1
        print(x)
        resultslist[timenames_list[z]] = framename
        z += 1
    return resultslist
        
the_output = make_day_df(timelist, timenamelist, july_trip,"pubTimeStamp")


the_output.keys()

the_output['2019-07-25-00-00']
# the_output['2019-05-01-00-00']['sumdid'].nunique()

the_output['2019-07-25-00-00']['sumdID'].value_counts()
# the_output['2019-05-01-00-00']['sumdid'].nunique()

the_output['2019-07-25-00-00']['sumdID'].value_counts().loc[lambda x : x<3]

35% had 3 or more

the_output['2019-07-25-00-00']['sumdID'].value_counts().loc[lambda x : x>5]

6% performed well with at least 6 rides that day.

#make a list of datetimes with datetime function 
start = datetime(2019, 7, 10, 0, 0)
end = datetime(2019, 7, 11, 0, 0)
delta = timedelta(days=1)
timelist = []
timenamelist = []
while start <= end:
    timelist.append(start)
    timenamelist.append(start.strftime("%Y-%m-%d-%H-%M"))
    start += delta
# print(timelist)
# print(timenamelist)

#goal of function: using list of time frames, in this case, days, make a dataframe containing
#info about the location for each scooter, for each day


#make dictionary of dataframes using base dataframe and a user-defined segments_of_time_list
#from cell above
def make_day_df(datetime_list, timenames_list, dataframe, colname):
    resultslist = {} #turn into dictionary
    x = 0
    y = 1
    z = 0
    while x in range(0,len(datetime_list)-1):
        starttime = datetime_list[x]
        endtime = datetime_list[y]
        mask = (dataframe[colname] > starttime) & (dataframe[colname] <= endtime)
        framename = dataframe.loc[mask]
        x += 1
        y += 1
        print(x)
        resultslist[timenames_list[z]] = framename
        z += 1
    return resultslist
        
the_output = make_day_df(timelist, timenamelist, july_trip,"pubTimeStamp")


the_output.keys()

the_output['2019-07-10-00-00']
# the_output['2019-05-01-00-00']['sumdid'].nunique()

the_output['2019-07-10-00-00']['sumdID'].value_counts()
# the_output['2019-05-01-00-00']['sumdid'].nunique()

1267 scooters total

the_output['2019-07-10-00-00']['sumdID'].value_counts().loc[lambda x : x<3]

269 had 3 or more. That's 21%

the_output['2019-07-10-00-00']['sumdID'].value_counts().loc[lambda x : x>5]

26 scooters had 6 or more rides, or 2%
